Change E to E(x0) and add __init__.py and change the name of the original file to lj_potential.py
